backend/requirements.txt
fastapi
uvicorn
pandas
chromadb
google-genai
python-dotenv